social_media_api/requirements.txt
asgiref==3.10.0
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
Django==5.2.8
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
